// src/_old-code/types/settings.types.ts
export type OpenMode = "active" | "tab" | "split" | "window";
// ...
export interface WriteDaily {
  type: "day";
}

export interface WriteWeekly {
  type: "week";
}

export interface WriteMonthly {
  type: "month";
}

export interface WriteQuarterly {
  type: "quarter";
}

export interface WriteYearly {
  type: "year";
}

export interface WriteWeekdays {
  type: "weekdays";
  weekdays: number[];
}

export interface WriteCustom {
  type: "custom";
  every: "day" | "week" | "month" | "quarter" | "year";
  duration: number;
  anchorDate: JournalAnchorDate;
}

export interface EndWritingNever {
  type: "never";
}

export interface EndWritingDate {
  type: "date";
  date: string;
}

export interface EndWritingAfterNTimes {
  type: "repeats";
  repeats: number;
}

export type FixedWriteIntervals = WriteDaily | WriteWeekly | WriteMonthly | WriteQuarterly | WriteYearly;

export interface JournalSettings {
  name: string;
  shelves: string[];

  write: FixedWriteIntervals | WriteCustom;

  confirmCreation: boolean;
  autoCreate: boolean;

  nameTemplate: string;
  dateFormat: string;
  folder: string;
  templates: string[];

  start: string;

  end: EndWritingNever | EndWritingDate | EndWritingAfterNTimes;

  index: {
    enabled: boolean;
    anchorDate: string;
    anchorIndex: number;
    allowBefore: boolean;
    type: "increment" | "reset_after";
    resetAfter: number;
  };

  commands: JournalCommand[];

  decorations: JournalDecoration[];

  navBlock: {
    type: "create" | "existing";
    rows: NavBlockRow[];
    decorateWholeBlock: boolean;
  };
  calendarViewBlock: {
    rows: NavBlockRow[];
    decorateWholeBlock: boolean;
  };

  frontmatter: {
    dateField: string;
    addStartDate: boolean;
    startDateField: string;
    addEndDate: boolean;
    endDateField: string;
    indexField: string;
  };
}

export interface PluginCommand {
  icon: string;
  name: string;
  writeType: FixedWriteIntervals["type"];
  type: "same" | "next" | "previous";
  showInRibbon: boolean;
  openMode: OpenMode;
}

export interface JournalCommand {
  icon: string;
  name: string;
  type:
    | "same"
    | "next"
    | "previous"
    | "same_next_week"
    | "same_previous_week"
    | "same_next_month"
    | "same_previous_month"
    | "same_next_year"
    | "same_previous_year";
  context: "today" | "open_note" | "only_open_note";
  showInRibbon: boolean;
  openMode: OpenMode;
}

export interface JournalDecoration {
  mode: "or" | "and";
  conditions: JournalDecorationCondition[];
  styles: JournalDecorationsStyle[];
}
// ...
export type JournalDecorationsStyle =
  | JournalDecorationBackground
  | JournalDecorationColor
  | JournalDecorationBorder
  | JournalDecorationShape
  | JournalDecorationCorner
  | JournalDecorationIcon;
// ...
export type JournalDecorationCondition =
  | JournalDecorationTitleCondition
  | JournalDecorationTagCondition
  | JournalDecorationPropertyCondition
  | JournalDecorationDateCondition
  | JournalDecorationWeekdayCondition
  | JournalDecorationOffsetCondition
  | JournalDecorationHasNoteCondition
  | JournalDecorationHasOpenTaskCondition
  | JournalDecorationAllTasksCompletedCondition;

export type NotesProcessing = "keep" | "clear" | "delete";

export interface NavBlockRow {
  template: string;
  fontSize: number;
  bold: boolean;
  italic: boolean;
  color: ColorSettings;
  background: ColorSettings;
  link: "none" | "self" | "journal" | Exclude<JournalSettings["write"]["type"], "custom">;
  journal: string;
  addDecorations: boolean;
}

export type ColorSettings =
  | {
      type: "transparent";
    }
  | {
      type: "theme";
      name: string;
    }
  | {
      type: "custom";
      color: string;
    };